requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.17
# Vulnerable version: mysql-connector-python==8.0.17
# Fixed version (CVE-2019-16782): mysql-connector-python==8.0.18